[Entity Store] Rename crud_client to crud
