+++ x-pack/solutions/security/plugins/entity_store/server/domain/logs_extraction/ccs_logs_extraction_client.ts
@@ -19,7 +19,7 @@ import {
   extractPaginationParams,
 } from './logs_extraction_query_builder';
 import { executeEsqlQuery, esqlResponseToBulkObjects } from '../../infra/elasticsearch/esql';
-import type { CRUDClient } from '../crud_client';
+import type { CRUDClient } from '../crud';
 
 export interface CcsExtractToUpdatesParams {
   type: EntityType;
