+++ x-pack/solutions/security/plugins/entity_store/server/request_context_factory.ts
@@ -21,7 +21,7 @@ import {
 } from './domain/definitions/saved_objects';
 import { CcsLogsExtractionClient, LogsExtractionClient } from './domain/logs_extraction';
 import { HistorySnapshotClient } from './domain/history_snapshot';
-import { CRUDClient } from './domain/crud_client';
+import { CRUDClient } from './domain/crud';
 import type { TelemetryReporter } from './telemetry/events';
 
 interface EntityStoreApiRequestHandlerContextDeps {
